% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarity.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Determines cosine similarity between two \code{\link{matrix}} objects}
\usage{
cosine_similarity(mat1 = NULL, mat2 = NULL)
}
\arguments{
\item{mat1}{\code{\link{matrix}} with row and column names}

\item{mat2}{\code{\link{matrix}} with row and column names}
}
\value{
\code{\link{matrix}} of cosine similarity between the input matrices
}
\description{
Determines the cosine similarity for two matrices. Requires that a dimension of the two input matrices is equal
}
\examples{

# cosine_similarity(my_matrix1, my_matrix2)

}
