% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cosine_similarity.R
\name{example_inputs}
\alias{example_inputs}
\title{shows example matrices for input}
\usage{
example_inputs()
}
\value{
list of example input matrices
}
\description{
returns a list of 2 matrices to unlist and view for example input
}
\examples{

# example_inputs()[1]
# example_inputs()[2]

}
