% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cosine_similarity.R
\name{test_cosine_similarity}
\alias{test_cosine_similarity}
\title{shows example output of cosine similarity}
\usage{
test_cosine_similarity()
}
\value{
matrix of cosine similarity between dimensions of input matrices
}
\description{
determines the cosine similarity for two matrices
}
\examples{

# test_cosine_similarity()

}
